requirements_full.txt
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
apscheduler>=3.10.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.97.0
